requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
coverage==7.10.7
Django==4.2.25
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
filelock==3.19.1
fsspec==2025.9.0
google-auth==2.41.1
google-genai==1.42.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
llvmlite==0.43.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20250625
packaging==25.0
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tomli==2.3.0
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yt-dlp==2025.9.26
